POLIO-1471: update permissions for rounds endpoint
- create PolioReadPermission class
- use it for RoundDashBoardViewset
- use it to replace existing class for LQASCountryViewset

Having more restricted permissions than campaigns API to avoid the mess of filtering by account for anon users and because this endpoint doesn't need to be public like the main campaigns API (which is used for the embedded calendar)
